requirements.txt
streamlit
opencv-python
torch
torchvision
transformers
Pillow
ultralytics
